Add files via upload
This study investigates the factors influencing livelihood transitions among rural-urban migrants in Guwahati, focusing on three key sectors: Construction and Transportation, as well as Hotels and Restaurants. The selected sectors encompass vital segments of migrant labor and different views on the movement flows and their causes. Knowing the mechanisms of livelihood transition is important for the legislators and stakeholders to face the problems of migrants sensitively and adequately.